binutils: ARC: Fix for variables in .tbss
Without this fix all variables that fall into .tbss section
(i.e. non-initialized thread local variables or those initialized with 0)
will be mapped to the same one memory location in run-time.

It was discovered while bringing-up Systemd on ARC because
thread local variables (and quite a few of them are initialized with 0)
are used in SystemD as a safe version of static variables.

Signed-off-by: Alexey Brodkin <abrodkin@synopsys.com>